+++ CursoWizelineCriptomonedas/app/src/main/java/com/vero/cursowizelinecriptomonedas/cryptoDetail/CryptoOrderListViewModel.kt
@@ -5,14 +5,18 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.vero.cursowizelinecriptomonedas.api.ApiResponseStatus
-import com.vero.cursowizelinecriptomonedas.cryptoInfoDetail.CryptoBookRepository
 import com.vero.cursowizelinecriptomonedas.model.CryptoBookDetail
 import com.vero.cursowizelinecriptomonedas.model.CryptoOrder
+import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.schedulers.Schedulers
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class CryptoOrderListViewModel : ViewModel() {
+@HiltViewModel
+class CryptoOrderListViewModel @Inject constructor(
+    private val cryptoOrderRepository: CryptoOrderRepository
+) : ViewModel() {
     private val _cryptoOrderList = MutableLiveData<List<CryptoOrder>>()
 
     val cryptoOrderList: LiveData<List<CryptoOrder>>
@@ -28,8 +32,6 @@ class CryptoOrderListViewModel : ViewModel() {
     val bookDetail: LiveData<CryptoBookDetail>
         get() = _bookDetail
 
-    private val cryptoOrderRepository = CryptoOrderRepository()
-    private val cryptoBookRepository = CryptoBookRepository()
 
     fun downloadCryptoOrder(crypto: String) {
         viewModelScope.launch {
@@ -45,13 +47,11 @@ class CryptoOrderListViewModel : ViewModel() {
         _status.value = apiResponseStatus
     }
 
-    fun downloadCryptoImage(crypto: String): String =
-        "https://firebasestorage.googleapis.com/v0/b/crypto-d6420.appspot.com/o/cryptocurrency_icon%2Fic_crypto_${
-            crypto.split("_").get(0)
-        }.png?alt=media"
+    fun downloadCryptoImage(crypto: String): String = cryptoOrderRepository.getCryptoImg(crypto)
+
 
     fun downloadCryptoBookDetail(crypto: String) {
-        cryptoBookRepository.getDetailBook(crypto)
+        cryptoOrderRepository.getDetailBook(crypto)
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe { onSuccess, onError ->
@@ -63,7 +63,7 @@ class CryptoOrderListViewModel : ViewModel() {
                     }
                 }
                 onError.let {
-                    
+
                 }
             }
     }
