+++ CursoWizelineCriptomonedas/app/src/main/res/layout/activity_crypto_detail.xml
@@ -8,6 +8,7 @@
         <variable
             name="crypto"
             type="com.vero.cursowizelinecriptomonedas.model.Crypto" />
+
         <variable
             name="cryptoBook"
             type="com.vero.cursowizelinecriptomonedas.model.CryptoBookDetail" />
